--- advent_of_code/aoc/2022/inputs/day12_sample.txt
@@ -1,5 +1,0 @@
-Sabqponm
-abcryxxl
-accszExk
-acctuvwj
-abdefghi
